requirements.txt
django
bs4
django-allauth
django-cors-headers
djangorestframework
dodopayments
drf-yasg
facebook-scraper
google-auth
instagrapi
pillow
python-dotenv
pyVirtualDisplay
selenium
webdriver-manager
dj_rest_auth
jwt
twikit
dnspython